service: ProductService

+++ E-Commerce/src/main/java/com/springproject/ecommerce/service/ProductService.java
@@ -18,7 +18,7 @@ public interface ProductService {
 	
 	List<Product> findAllProductsForAdmin();
 
-	List<Product> findProductByCategoryId(Integer categoryId);
+	List<Product> findProductByCategoryId(Long categoryId);
 
 	Product findProductByIdForAdmin(int id);
